+++ src/votekit/cvr_loaders.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pathlib
 import os
 from typing import Optional
-from .profile import PreferenceProfile
+from .pref_profile import PreferenceProfile
 from .ballot import Ballot
 from pandas.errors import EmptyDataError, DataError
 from fractions import Fraction
