+++ src/votekit/graphs/pairwise_comparison_graph.py
@@ -4,7 +4,7 @@ import networkx as nx  # type: ignore
 import matplotlib.pyplot as plt  # type: ignore
 
 from ..ballot import Ballot
-from ..profile import PreferenceProfile
+from ..pref_profile import PreferenceProfile
 from .models import Graph
