+++ src/votekit/models.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import Union, Any
-from .profile import PreferenceProfile
+from .pref_profile import PreferenceProfile
 from .election_state import ElectionState
 from .ballot_generator import BallotGenerator
