+++ tests/test_cleaning.py
@@ -1,4 +1,4 @@
-from votekit.profile import PreferenceProfile
+from votekit.pref_profile import PreferenceProfile
 from votekit.ballot import Ballot
 from votekit.cleaning import remove_empty_ballots, deduplicate_profiles, remove_noncands
 from fractions import Fraction
