+++ tests/test_utils.py
@@ -1,5 +1,5 @@
 from votekit.utils import mentions, first_place_votes, borda_scores
-from votekit.profile import PreferenceProfile
+from votekit.pref_profile import PreferenceProfile
 from votekit.ballot import Ballot
 from fractions import Fraction
